Allow use of IAM Role for Service Account -  IRSA (#194)

+++ tools/backup/backup.sh
@@ -244,19 +244,22 @@ function activate_gcp() {
 }
 
 function activate_aws() {
-  echo "Activating aws credentials before beginning"
-  mkdir -p /root/.aws/
-  cp /credentials/credentials ~/.aws/config
-
-  if [ $? -ne 0 ]; then
-    echo "Credentials failed; no way to copy to aws."
-    exit 1
-  fi
-
-  aws sts get-caller-identity
-  if [ $? -ne 0 ]; then
-    echo "Credentials failed; no way to copy to aws."
-    exit 1
+  local credentials="/credentials/credentials"
+  if [[ -f "${credentials}" ]]; then
+    echo "Activating aws credentials before beginning"
+    mkdir -p /root/.aws/
+    cp /credentials/credentials ~/.aws/config
+    if [ $? -ne 0 ]; then
+      echo "Credentials failed; no way to copy to aws."
+      exit 1
+    fi
+    aws sts get-caller-identity
+    if [ $? -ne 0 ]; then
+      echo "Credentials failed; no way to copy to aws."
+      exit 1
+    fi
+  else
+    echo "No credentials file found. Assuming IAM Role for Service Account - IRSA is configured"
   fi
 }
